refactor of resilience config
