+++ conf/application.conf
@@ -207,4 +207,6 @@ resilience {
 
   # maximum tasks per resilient function to process per tick
   max-tasks = 1
+
+  des-enabled-on = 2020-04-06
 }
